bigquery_table_manager/requirements.txt
configparser==4.0.2
google-cloud-bigquery==1.23.1
protobuf~=3.20.0
python-dateutil==2.8.2